+++ src/ts/util/editorCommonEvent.ts
@@ -75,6 +75,9 @@ export const cutEvent =
     };
 
 export const scrollCenter = (vditor: IVditor) => {
+    if (vditor.currentMode === "wysiwyg" && vditor.options.comment.enable) {
+        vditor.options.comment.adjustTop(vditor.wysiwyg.getComments(vditor, true))
+    }
     if (!vditor.options.typewriterMode) {
         return;
     }